+++ R/CalExp2dCt.R
@@ -9,11 +9,11 @@ globalVariables(c(
 #' Calculate relative gene expression using the 2-dCt method
 #' with a reference gene for normalization.
 #'
-#' @param cq.table A data frame containing position and Cq values.
+#' @param cq_table A data frame containing position and Cq values.
 #'   Must have columns: Position, Gene, Cq.
-#' @param design.table A data frame containing position and group information.
+#' @param design_table A data frame containing position and group information.
 #'   Must have columns: Position, Group, BioRep.
-#' @param ref.gene Character. The name of the reference gene (default: "Actin").
+#' @param ref_gene Character. The name of the reference gene (default: "Actin").
 #'
 #' @return A data frame with expression values, including columns:
 #'   position, cq, group, gene, biorep, mean.cq, expre, n, mean.expre,
@@ -28,26 +28,26 @@ globalVariables(c(
 #' \dontrun{
 #' df1.path <- system.file("examples", "dct.cq.txt", package = "qPCRtools")
 #' df2.path <- system.file("examples", "dct.design.txt", package = "qPCRtools")
-#' cq.table <- read.table(df1.path, sep = ",", header = TRUE)
-#' design.table <- read.table(df2.path, sep = ",", header = TRUE)
-#' res <- CalExp2dCt(cq.table, design.table, ref.gene = "Actin")
+#' cq_table <- read.table(df1.path, sep = ",", header = TRUE)
+#' design_table <- read.table(df2.path, sep = ",", header = TRUE)
+#' res <- CalExp2dCt(cq_table, design_table, ref_gene = "Actin")
 #' head(res)
 #' }
 #'
 #' @author Xiang LI <lixiang117423@gmail.com>
-CalExp2dCt <- function(cq.table,
-                       design.table,
-                       ref.gene = "Actin") {
-  if (!is.data.frame(cq.table)) {
-    stop("'cq.table' must be a data frame")
+CalExp2dCt <- function(cq_table,
+                       design_table,
+                       ref_gene = "Actin") {
+  if (!is.data.frame(cq_table)) {
+    stop("'cq_table' must be a data frame")
   }
-  if (!is.data.frame(design.table)) {
-    stop("'design.table' must be a data frame")
+  if (!is.data.frame(design_table)) {
+    stop("'design_table' must be a data frame")
   }
 
   # merge data
-  cq.table %>%
-    dplyr::left_join(design.table, by = "Position") %>%
+  cq_table %>%
+    dplyr::left_join(design_table, by = "Position") %>%
     dplyr::rename(
       position = Position,
       cq = Cq,
@@ -58,7 +58,7 @@ CalExp2dCt <- function(cq.table,
 
   # reference gene
   df %>%
-    dplyr::filter(gene == ref.gene) %>%
+    dplyr::filter(gene == ref_gene) %>%
     dplyr::group_by(group, biorep) %>%
     dplyr::mutate(
       mean.cq = mean(cq),
@@ -70,7 +70,7 @@ CalExp2dCt <- function(cq.table,
 
   # target gene
   df %>%
-    dplyr::filter(gene != ref.gene) %>%
+    dplyr::filter(gene != ref_gene) %>%
     dplyr::mutate(temp = paste0(group, biorep)) %>%
     dplyr::left_join(df.ref, by = "temp") %>%
     dplyr::select(-temp) %>%